2026-review/pyproject.toml
[project]
name = "2026-review"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=12.2.0",
    "requests>=2.33.1",
]
